pyproject.toml
[build-system]
requires = ["setuptools >= 60.0"]
build-backend = "setuptools.build_meta"

[project]
name = "xasNormalisation"
version = "1.0.0"
requires-python = ">= 3.10"
dependencies = ["pandas", "matplotlib", "scipy","xraylarch"]
authors = [{name = "Kenneth P. Marshall", email = "kenneth.marshall@esrf.fr"}]
maintainers = [{name = "Kenneth P. Marshall", email = "kenneth.marshall@esrf.fr"}]
description = "XAS COlumn selection, Regridding and Normalisation. Programs for splitting files, selecting columns, regridding and normalising XAS data on BM31"
readme = {file = "README.MD", content-type = "text/markdown"}
keywords = ["XAS","X-ray absorption spectroscopy"]
classifiers = [
"Development Status :: 4 - Beta",
"Programming Language :: Python :: 3.12"]

[project.scripts]
xasCorn = "xasCorn.xasColumnsNormBkg:main"
h5columns = "xasCorn.h5Processor:runLoop"

[tool.setuptools]
packages = ["xasCorn"]
include-package-data = true

[tool.setuptools.package-data]
edgeList = ["xasCorn/edgeList.txt"]

[project.urls]
repository = "https://github.com/msujas/xas_normalisation"
